refactor yaml io to kio
