--- internal/yaml/doc.go
@@ -1,2 +1,0 @@
-// Package yaml contains utilities for working with YAML manifests.
-package yaml
