+++ pkg/cluster/loader.go
@@ -1,20 +1,13 @@
 package cluster
 
 import (
-	"os"
-
-	"gopkg.in/yaml.v3"
+	"github.com/go-kure/kure/pkg/kio"
 )
 
 // LoadClusterFromYAML reads and parses a cluster configuration from a YAML file.
 func LoadClusterFromYAML(configPath string) (*Cluster, error) {
-	data, err := os.ReadFile(configPath)
-	if err != nil {
-		return nil, err
-	}
-
 	var c Cluster
-	if err := yaml.Unmarshal(data, &c); err != nil {
+	if err := kio.LoadFile(configPath, &c); err != nil {
 		return nil, err
 	}
 	return &c, nil
