+++ pkg/fluxcd/bootstrap.go
@@ -1,13 +1,13 @@
 package fluxcd
 
 import (
-	"os"
 	"time"
 
 	sourcev1beta2 "github.com/fluxcd/source-controller/api/v1beta2"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/yaml"
+
+	"github.com/go-kure/kure/pkg/kio"
 )
 
 // NewOCIRepositoryYAML constructs an OCIRepository resource from config.
@@ -31,22 +31,13 @@ func NewOCIRepositoryYAML(cfg *OCIRepositoryConfig) *sourcev1beta2.OCIRepository
 
 // WriteYAMLResource marshals the object to YAML at the given path.
 func WriteYAMLResource(path string, obj client.Object) error {
-	data, err := yaml.Marshal(obj)
-	if err != nil {
-		return err
-	}
-	return os.WriteFile(path, data, 0644)
+	return kio.SaveFile(path, obj)
 }
 
 // PatchOCIRepositoryFromFile loads an OCIRepository from path and applies a patch function.
 func PatchOCIRepositoryFromFile(path string, patchFn func(*sourcev1beta2.OCIRepository) error) error {
-	raw, err := os.ReadFile(path)
-	if err != nil {
-		return err
-	}
-
 	var repo sourcev1beta2.OCIRepository
-	if err := yaml.Unmarshal(raw, &repo); err != nil {
+	if err := kio.LoadFile(path, &repo); err != nil {
 		return err
 	}
 
@@ -54,12 +45,7 @@ func PatchOCIRepositoryFromFile(path string, patchFn func(*sourcev1beta2.OCIRepo
 		return err
 	}
 
-	newData, err := yaml.Marshal(repo)
-	if err != nil {
-		return err
-	}
-
-	return os.WriteFile(path, newData, 0644)
+	return kio.SaveFile(path, repo)
 }
 
 func parseDurationOrDefault(s string) time.Duration {
