+++ pkg/kio/runtime_test.go
@@ -1,4 +1,4 @@
-package yaml
+package kio
 
 import (
 	"errors"
